Move default functions to helpers
