+++ src/sentry/api/endpoints/notification_defaults.py
@@ -4,40 +4,8 @@ from rest_framework.response import Response
 from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import Endpoint, control_silo_endpoint
-from sentry.notifications.defaults import (
-    NOTIFICATION_SETTING_DEFAULTS,
-    NOTIFICATION_SETTINGS_ALL_SOMETIMES,
-    NOTIFICATION_SETTINGS_ALL_SOMETIMES_V2,
-)
-from sentry.notifications.types import (
-    NOTIFICATION_SETTING_OPTION_VALUES,
-    NOTIFICATION_SETTING_TYPES,
-)
+from sentry.notifications.helpers import get_provider_defaults, get_type_defaults
 from sentry.types.integrations import EXTERNAL_PROVIDERS
-
-
-def get_provider_defaults():
-    # create the data structure outside the endpoint
-    provider_defaults = []
-    for key, value in NOTIFICATION_SETTING_DEFAULTS.items():
-        provider = EXTERNAL_PROVIDERS[key]
-        # if the value is NOTIFICATION_SETTINGS_ALL_SOMETIMES then it means the provider
-        # is on by default
-        if value == NOTIFICATION_SETTINGS_ALL_SOMETIMES:
-            provider_defaults.append(provider)
-    return provider_defaults
-
-
-def get_type_defaults():
-    # this tells us what the default value is for each notification type
-    type_defaults = {}
-    for key, value in NOTIFICATION_SETTINGS_ALL_SOMETIMES_V2.items():
-        # for the given notification type, figure out what the default value is
-        notification_type = NOTIFICATION_SETTING_TYPES[key]
-        default = NOTIFICATION_SETTING_OPTION_VALUES[value]
-        type_defaults[notification_type] = default
-    return type_defaults
-
 
 PROVIDER_DEFAULTS = get_provider_defaults()
 TYPE_DEFAULTS = get_type_defaults()
